BOth Chat and Indox feature comleted
