+++ src/main/java/com/TextIt/model/utils/CommonMethods.java
@@ -17,6 +17,10 @@ public class CommonMethods {
     public static final String PURPLE = "\u001B[35m";
     public static final String CYAN = "\u001B[36m";
     public static final String BOLD = "\u001B[1m";
+
+    public static final String BG_RED = "\u001B[41m";
+
+    public static final String CLEAR_SCREEN = "\u001B[2J\u001B[H";
     //Objects
     static Scanner scanner = new Scanner(System.in);
     private static final DataBase db = new DataBase();
@@ -289,23 +293,35 @@ public class CommonMethods {
 
     public static int featchIdForNotification(String content , String type) throws SQLException {
         DataBase db = new DataBase();
-        Connection conn = DriverManager.getConnection(db.getUrl(),db.getUsername(), db.getPassword());
-        if(type.equalsIgnoreCase("message")) {
-            PreparedStatement ps = conn.prepareStatement("select id from messages where  message= ? ");
-            ps.setString(1, content);
-            ResultSet rs = ps.executeQuery();
-            rs.next();
-            return rs.getInt(1);
-        }else if (type.equalsIgnoreCase("Comment")) {
-            PreparedStatement ps = conn.prepareStatement("select c_id from comments where content = ? ");
-            ps.setString(1, content);
-            ResultSet rs = ps.executeQuery();
-            rs.next();
-            return rs.getInt(1);
-        }else{
-            System.out.println("Unsupported notification type: " + type);
+        try(Connection conn = DriverManager.getConnection(db.getUrl(),db.getUsername(), db.getPassword())) {
+            if (type.equalsIgnoreCase("message")) {
+                PreparedStatement ps = conn.prepareStatement("select id from messages where  message= ? ");
+                ps.setString(1, content);
+                ResultSet rs = ps.executeQuery();
+                rs.next();
+                return rs.getInt(1);
+            } else if (type.equalsIgnoreCase("Comments")) {
+                PreparedStatement ps = conn.prepareStatement("select c_id from comments where content = ? ");
+                ps.setString(1, content);
+                ResultSet rs = ps.executeQuery();
+                rs.next();
+                return rs.getInt(1);
+            } else if (type.equalsIgnoreCase("like")) {
+                PreparedStatement ps = conn.prepareStatement("select like_id from likes where user_id = ? ");
+                ps.setString(1, content);
+                ResultSet rs = ps.executeQuery();
+                rs.next();
+                return rs.getInt(1);
+
+            } else {
+                System.out.println("Unsupported notification type: " + type);
+                return -1;
+            }
+        }catch (SQLException e) {
+            e.printStackTrace();
             return -1;
         }
+
 
     }
